History and errors
Partial fix on history where if you have the page in your histroy it would previously have jsut redirected to main page due to being single site. hoping to do it to work off the back button.

Fixed issue where the left box wouldnt neccessarily wipe if loaded in a certain order, forgot to tell it to wipe if loading two left ones back to back